requirements.txt
llama-index==0.14.19
llama-index-core==0.14.19
llama-index-llms-openai==0.7.4
llama-index-embeddings-openai==0.6.0
llama-index-vector-stores-chroma==0.5.5
llama-index-workflows==2.17.1
openai==2.30.0
chromadb==1.5.5
python-dotenv==1.2.2
python-telegram-bot==22.7
pdfplumber==0.11.9
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==12.1.1
